requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
PyPDF2
langchain
